+++ lib/src/events_internal.dart
@@ -10,7 +10,7 @@ class EventsInternal implements Events {
   final CountlyState _countlyState;
 
   @override
-  Future<String?> recordEvent(String key, [Map<String, Object>? segmentation, int? count, int? sum, int? duration]) async {
+  Future<String?> recordEvent(String key, [Map<String, Object>? segmentation, int? count, double? sum, int? duration]) async {
     return _internalEventMethodCall(key, 'recordEvent', segmentation, count ?? 1, sum ?? 0, duration ?? 0);
   }
 
@@ -20,7 +20,7 @@ class EventsInternal implements Events {
   }
 
   @override
-  Future<String?> endEvent(String key, [Map<String, Object>? segmentation, int? count, int? sum]) async {
+  Future<String?> endEvent(String key, [Map<String, Object>? segmentation, int? count, double? sum]) async {
     return _internalEventMethodCall(key, 'endEvent', segmentation, count ?? 1, sum ?? 0);
   }
 
@@ -29,7 +29,7 @@ class EventsInternal implements Events {
     return _internalEventMethodCall(key, 'cancelEvent');
   }
 
-  Future<String?> _internalEventMethodCall(String key, String method, [Map<String, Object>? segmentation, int? count, int? sum, int? duration]) async {
+  Future<String?> _internalEventMethodCall(String key, String method, [Map<String, Object>? segmentation, int? count, double? sum, int? duration]) async {
     List<Object?> args = [];
 
     if (!_countlyState.isInitialized) {
